First Operating version of exercise loader.
Commit of the first operating version of lesson_loader. It's always a raw version with memory leak and esay bug.
It's modify the file's of lesson which are necassary to include the exercise in the lesson.
For now, it's only work with little exercise and with lesson which respect generic naming of file.